refactor: resizer component and remove div resizer
