+++ src/component/elements/resizer/SVGResizer.tsx
@@ -1,8 +1,8 @@
+/* eslint-disable react/no-unused-prop-types */
 /** @jsxImportSource @emotion/react */
 import { css } from '@emotion/react';
 import { CSSProperties } from 'react';
 
-import { ResizerProps } from './Resizer';
 import useResizer from './useResizer';
 
 const style: Record<'anchor' | 'innerContainer', CSSProperties> = {
@@ -32,6 +32,26 @@ const styles = {
     }
   `,
 };
+
+export interface Position {
+  x1: number;
+  x2: number;
+}
+
+type ChildType = React.ReactElement[] | React.ReactElement | boolean | null;
+
+export interface ResizerProps {
+  children?: ChildType | ((position: Position, isActive: boolean) => ChildType);
+  position: Position;
+  onStart?: PositionChangeHandler;
+  onMove?: PositionChangeHandler;
+  onEnd?: PositionChangeHandler;
+  parentElement?: HTMLElement | null;
+  dragHandleClassName?: string;
+  disabled?: boolean;
+}
+
+type PositionChangeHandler = (data: Position) => void;
 
 export default function SVGResizer(props: ResizerProps) {
   const { children, disabled, position } = props;
